backend/requirements.txt
azure-functions
azure-functions-durable
azure-communication-email
python-dotenv
requests
